Finished!
The files that we will be working on are located in the Production_Files folder. The other folders are just other assets I have used or tested that I decide to keep in case we decide to use them.

KriptoFX is the water engine folder, palms are the palm trees folder, and SubtleSand is the sand materials folder.

I have made a very, very basic terrain in BeachScene_Production Scene. We will be developing that scene for our project.